refactored the myArticles page
