+++ frontend/src/pages/myArticles/index.tsx
@@ -1,21 +1,19 @@
-import React from 'react';
 import { useParams } from 'react-router-dom';
 
 import { getUser } from '@utils/getUser';
-import { fetchWrapper } from '@utils/fetchWrapper';
 
 import { useScreenWidth } from '@hooks/useScreenWidth';
+import { useFetchData } from '@hooks/useFetchData';
 
-import { ArticleSmall } from '@components/articles/articleSmall';
-import { ArticlePrivate } from '@components/articles/articlePrivate';
 import { SkeletonCategoryPanel } from '@components/common/skeletons/skeletonCategoryPanel';
+import { Separator } from '@components/animation/separator';
+
+import { ArticleRender } from '@pages/myArticles/components/main/articleRender';
+import { CreateButton } from '@pages/myArticles/components/main/createButton';
 
 import { Box, Heading, Pagination } from '@primer/react';
-import { PencilIcon } from '@primer/octicons-react';
 
 export const MyArticles = () => {
-  const [publicArticles, setPublicArticles] = React.useState<any>(null);
-  const [privateArticles, setPrivateArticles] = React.useState<any>(null);
   const screenWidth = useScreenWidth();
   const { page } = useParams();
 
@@ -26,41 +24,21 @@ export const MyArticles = () => {
     return (window.location.href = '/');
   }
 
-  React.useEffect(() => {
-    const controller = new AbortController();
-    const signal = controller.signal;
+  const { data: publicArticles } = useFetchData<[]>(
+    `${backendUrl}/articles/author?authorName=${user?.Username}&page=${page}`,
+    [],
+    {},
+    true,
+    60 * 15
+  );
 
-    fetchWrapper(
-      `${backendUrl}/articles/author?authorName=${user?.Username}&page=${page}`,
-      { signal },
-      true,
-      60 * 15
-    ).then((data) => {
-      setPublicArticles(data.response.return);
-    });
-
-    fetchWrapper(
-      `${backendUrl}/articles/author?authorName=${user?.Username}&visibility=private`,
-      { signal },
-      true,
-      60 * 15
-    ).then((data) => {
-      setPrivateArticles(data.response.return);
-    });
-
-    return () => {
-      controller.abort();
-    };
-  }, []);
-
-  const handleCreate = () => {
-    if (user?.Verified == 'true') {
-      return (window.location.href = '/create');
-    }
-    alert(
-      'This account is not verified. Please verify your email to create articles.'
-    );
-  };
+  const { data: privateArticles } = useFetchData<[]>(
+    `${backendUrl}/articles/author?authorName=${user?.Username}&visibility=private`,
+    [],
+    {},
+    true,
+    60 * 15
+  );
 
   return (
     <Box
@@ -73,17 +51,14 @@ export const MyArticles = () => {
         mb: '100px',
       }}
     >
+      {/* Title */}
       <Heading sx={{ fontSize: screenWidth < 768 ? '28px' : '42px' }}>
         Your Articles
       </Heading>
-      <Box
-        sx={{
-          width: '100%',
-          height: '1px',
-          backgroundColor: 'ansi.black',
-        }}
-      ></Box>
 
+      <Separator />
+
+      {/* Article display */}
       <Box
         sx={{
           display: 'flex',
@@ -96,49 +71,19 @@ export const MyArticles = () => {
       >
         {publicArticles && privateArticles ? (
           <>
-            {publicArticles.map((item: any, index: any) => (
-              <ArticleSmall key={index} article={item} />
-            ))}
-            {privateArticles.map((item: any, index: any) => (
-              <ArticlePrivate key={index} article={item} />
-            ))}
-            <Box
-              onClick={handleCreate}
-              sx={{
-                width: '330px',
-                height: '260px',
-                position: 'relative',
-                borderRadius: '10px',
-                border: '1px solid',
-                borderColor: 'ansi.black',
-                boxShadow: '0px 0px 25px rgba(0, 255, 0, 0)',
-                transition: '0.3s all',
-                ':hover': {
-                  boxShadow: '0px 0px 15px rgba(0, 255, 0, 0.4)',
-                },
-              }}
-            >
-              <Box
-                sx={{
-                  width: '80px',
-                  position: 'absolute',
-                  left: '50%',
-                  top: '20%',
-                  transform: 'translate(-50%,0%)',
-                  display: 'flex',
-                  flexDirection: 'column',
-                  alignItems: 'center',
-                }}
-              >
-                <PencilIcon size={80} />
-                <Heading sx={{ cursor: 'default' }}>Create</Heading>
-              </Box>
-            </Box>
+            <ArticleRender
+              publicArticles={publicArticles}
+              privateArticles={privateArticles}
+            />
+
+            <CreateButton user={user} />
           </>
         ) : (
           <SkeletonCategoryPanel bigArticles={false} />
         )}
       </Box>
+
+      {/* Pagination buttons */}
       <Pagination
         currentPage={Number(page) || 1}
         pageCount={99}
